Cypress and report updated
