--- cypress.json
@@ -1,37 +1,0 @@
-{
-    "baseUrl": "http://automationpractice.com/index.php",
-    "defaultCommandTimeout" : 120000,
-    "viewportHeight": 768,
-    "viewportWidth": 1024,
-    "env": {
-        "accountPageUrl" : "http://automationpractice.com/index.php?controller=my-account",
-        "accountCreationUrl": "http://automationpractice.com/index.php?controller=authentication&back=my-account#account-creation",
-        "welcomeText": "Welcome to your account. Here you can manage all of your personal information and orders.",
-        "addedWishConfirmationMsg":"Added to your wishlist.",
-        "productAddedMsg": "Product successfully added to your shopping cart",        
-        "emptyCartMsg": "Your shopping cart is empty.",
-        "orderCompleteMsg": "Your order on My Store is complete.",
-        "finalPrice":"$29.00",
-        "paymentType": "Bank-wire payment.",
-        "moreThanHundred": "You may not request more than 100 items.",
-        "missingHashmsg": " You must provide a hash.",
-        "missingTimestamp": "You must provide a timestamp."
-    },
-    "reporter": "cypress-multi-reporters",
-    "reporterOptions":{
-        "reporterEnabled": "mochawesome",
-        "mochawesomeReporterOptions":{
-                "charts": true,
-                "reportDir": "cypress/reports/mocha",
-                "reportTitle": "Senior QA Challenge: Raffy Rodriguez's Test's results",
-                "reportFilename" : "report",
-                "reportPageTitle": "Senior QA Challenge's Test Results",
-                "quiet": true,                
-                "overwrite": false,
-                "html": false,
-                "json": true,
-                "timestamp": "mm/dd/yyyy HH:MM"
-    }
-    },
-    "video": false
-}
