--- cypress/integration/QAChallenge/TS_001 Login and Logout.spec.js
@@ -1,17 +1,0 @@
-import SignInPage from '../../support/pageObjets/signInPage';
-
-const signIn = new SignInPage();
-
-describe('Login and Logout', function(){
-
-  it('Login successfully', function(){
-    cy.login();
-    cy.url().should('eq', Cypress.env('accountPageUrl'));
-  });
-
-  it('Logout successfully', function(){
-    cy.signOut();
-    signIn.getAuthObject().should('be.visible');
-  });
-
-});
